major refactor of the Maps class into mixin classes
